% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addOntology.R
\name{addOntology}
\alias{addOntology}
\title{cleanTraitNames function}
\usage{
addOntology(testData, ontology)
}
\arguments{
\item{testData}{[value]}

\item{ontology}{[value]. Default is TRUE}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
# none
}
