% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addPlusMinus.R
\name{addPlusMinus}
\alias{addPlusMinus}
\title{addPlusMinus function}
\usage{
addPlusMinus(smry, traits = NULL)
}
\arguments{
\item{smry}{summary output from oneYearOneLocSummary, oneYearOverLocSummary, or multiYearSummary}

\item{traits}{list of traits to provide LSD from mean to}
}
\value{
[value]
}
\description{
function to add LSD separation from the mean
}
\details{
[fill in details here]
}
\examples{
# none
}
