% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addTraitPercRank.R
\name{addTraitPercRank}
\alias{addTraitPercRank}
\title{formatPrintSummary function}
\usage{
addTraitPercRank(
  smry,
  trait,
  keep = "Line",
  calcPercMean = TRUE,
  sign = -1,
  ...
)
}
\arguments{
\item{smry}{summary output from oneYearOneLocSummary, oneYearOverLocSummary, or multiYearSummary}

\item{trait}{trait to calculate percent of mean and rank}

\item{keep}{columns to keep at begining of data frame, PercMean and Rank will be added to beginning of data.frame}

\item{calcPercMean}{should PercMean be calculated?}
}
\value{
[value]
}
\description{
function to sort summaries.
}
\details{
[fill in details here]
}
\examples{
# none
}
