% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitConversions.R
\name{awnsToScore}
\alias{awnsToScore}
\title{awnsToScore function}
\usage{
awnsToScore(awns)
}
\arguments{
\item{awns}{[value]}
}
\value{
[value]
}
\description{
function to convert awns A, TA, AL) to an awn score (1, 2, 3)
}
\details{
[fill in details here]
}
\examples{
none
}
