% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betterImage.R
\name{betterImage}
\alias{betterImage}
\title{mergreNotes function
betterImage function}
\usage{
betterImage(X, ...)
}
\arguments{
\item{X}{matrix of values to be passed to image()}
}
\value{
plot image in correct orientation
}
\description{
function to draw images in the same orientation as the matrix
}
\details{
[fill in details here]
}
\examples{
betterImage(volcano)
}
