% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLUEindex.R
\name{BLUEindex}
\alias{BLUEindex}
\title{roundSummary function}
\usage{
BLUEindex(envMeans, cv, weights = 1, scaleb = TRUE)
}
\arguments{
\item{envMeans}{matrix of environment means}

\item{cv}{coefficicent of variation}

\item{weights}{weights for each environment. default is 1/ncol(envMeans)}

\item{scaleb}{should the index be scaled to 1? default is TRUE.}
}
\value{
[value]
}
\description{
function to round summaries. Only works if plus minus present
}
\details{
[fill in details here]
}
\examples{
none
}
