% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitConversions.R
\name{buacToKgha}
\alias{buacToKgha}
\title{buacToKgha function}
\usage{
buacToKgha(buac, barley = FALSE)
}
\arguments{
\item{barley}{logical. Is it barley?}

\item{bu}{numeric. Bushels.}
}
\value{
[value]
}
\description{
function to convert bushels per acre to kilograms per hectare based on standard testweight (60 for wheat and 48 for barley)
}
\details{
[fill in details here]
}
\examples{
# none
}
