% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkPlotNos.R
\name{checkPlotNos}
\alias{checkPlotNos}
\title{checkPlotNos function}
\usage{
checkPlotNos(x, type, testName = NULL)
}
\arguments{
\item{x}{[value]}

\item{type}{[value]}

\item{testName}{[value]. Default is NULL}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
