% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanAwns.R
\name{cleanAwns}
\alias{cleanAwns}
\title{selectAwn function}
\usage{
cleanAwns(x)
}
\arguments{
\item{x}{vector of multiple awn observations of the form "WARVA: A,TA,A"}
}
\value{
[value]
}
\description{
function to select an awn type from multiple observations. Intended to be used from a call from cleanAwns(). TA/AL conflicts are assumed to be TA, while TA/A  are assumed to be awned. A/AL conflicts will be printed as A,AL
}
\details{
[fill in details here]
}
\examples{
# none
}
