% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanDJ.R
\name{cleanDJ}
\alias{cleanDJ}
\title{cleanDJ function}
\usage{
cleanDJ(djdf, year, testName = NULL, minTW = 55, maxTW = 64,
  moistThresh = 16, moistTooHighNA = TRUE, outsideTWrangeNA = FALSE,
  rmChars = FALSE, barley = FALSE)
}
\arguments{
\item{djdf}{[value]}

\item{year}{[value]}

\item{testName}{[value]. Default is NULL}

\item{minTW}{[value]. Default is 56}

\item{maxTW}{[value]. Default is 63}

\item{moistTooHighNA}{[value]. Default is TRUE}

\item{outsideTWrangeNA}{[value]. Default is FALSE}

\item{rmChars}{[value]. Should plot numbers with characters [A-z] be removed? Default is FALSE}

\item{barley}{[value]. Default is FALSE}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
# none
}
