% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanHarvM.R
\name{cleanHarvM}
\alias{cleanHarvM}
\title{cleanHarvM function}
\usage{
cleanHarvM(hmi, lbs = TRUE, forgotCycleThreshold = 0.2, qnt = 0.99,
  keepCols = NULL)
}
\arguments{
\item{hmi}{[value]}

\item{lbs}{[value]. Default is TRUE}

\item{forgotCycleThreshold}{[value]. Default is 0.2}

\item{qnt}{[value]. Default is 0.99}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
# none
}
