% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanScale.R
\name{cleanScale}
\alias{cleanScale}
\title{cleanScale function}
\usage{
cleanScale(scl, year, testName = NULL, tare = NULL, minWt = 500,
  maxWt = 6000, rmDup = FALSE)
}
\arguments{
\item{scl}{[value]}

\item{year}{[value]}

\item{testName}{[value]. Default is NULL}

\item{tare}{[value]. Default is NULL}

\item{minWt}{[value]. Default is 500}

\item{maxWt}{[value]. Default is 6000}

\item{rmDup}{[value]. Default is FALSE}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
# none
}
