% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanScores.R
\name{cleanScores}
\alias{cleanScores}
\title{cleanScores function}
\usage{
cleanScores(fb, scoreTraits = NULL)
}
\arguments{
\item{fb}{[value]}

\item{scoreTraits}{[value]. Default is NULL}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
