% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanTraitNames.R
\name{cleanTraitNames}
\alias{cleanTraitNames}
\title{cleanTraitNames function}
\usage{
cleanTraitNames(traitNames, addUnits = TRUE)
}
\arguments{
\item{traitNames}{[value]}

\item{addUnits}{[value]. Default is TRUE}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
