% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitConversions.R
\name{cmToInch}
\alias{cmToInch}
\title{cmToInch function}
\usage{
cmToInch(cm)
}
\arguments{
\item{cm}{[value]}
}
\value{
[value]
}
\description{
function to convert centimeters to inches.
}
\details{
[fill in details here]
}
\examples{
# none
}
