% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compileNotes.R
\name{compileNotes}
\alias{compileNotes}
\title{compileNotes function}
\usage{
compileNotes(dfc, charTrtCols)
}
\arguments{
\item{dfc}{data.frame with}

\item{charTrtCols}{character vector with column numbers as elements and names(charTrtCols) as names of traits}
}
\value{
line entry trial and contactenated notes across plots for each
}
\description{
function to compile notes taken across plots and trials for each line.
}
\details{
[fill in details here]
}
\examples{
none
}
