% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensusTW.R
\name{consensusTW}
\alias{consensusTW}
\title{consensusTW function}
\usage{
consensusTW(djhm, defaultDJ = FALSE, TWrange = c(55, 64),
  barley = FALSE, ...)
}
\arguments{
\item{djhm}{data.frame with 'Moisture', 'TestWeight' from th Dickey-John, and 'Moisture_HarvM' and 'TestWeight_HarvM' from the harvest master.}

\item{defaultDJ}{logical. Should default values of moisture and test weight be from the dickey john? Harvest master otherwise.}

\item{TWrange}{numeric. vector of length 2 to determeine valid test weights. default is c(55, 64)}

\item{barley}{logical. Is barley?}
}
\value{
[value]
}
\description{
function to create a consensus test weight and moisture
}
\details{
[fill in details here]
}
\examples{
# none
}
