% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensusYld.R
\name{consensusYld}
\alias{consensusYld}
\title{consensusYld function}
\usage{
consensusYld(
  sclhm,
  defaultScale = TRUE,
  yldRange = c(500, 6000),
  barley = FALSE,
  ...
)
}
\arguments{
\item{sclhm}{data.frame with 'netWieght' from a scale, and 'netWeight_HarvM' from the harvest master.}

\item{defaultScale}{logical. Should default values of netweight be from the scale? Harvest master otherwise.}

\item{yldRange}{numeric. vector of length 2 to determeine valid netweights. default is c(500, 5000)}

\item{barley}{logical. Is barley?}
}
\value{
[value]
}
\description{
function to create a consensus yld measurement from Harvest master and scale data
}
\details{
[fill in details here]
}
\examples{
# none
}
