% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitConversions.R
\name{correctWeight}
\alias{correctWeight}
\title{correctWeight function}
\usage{
correctWeight(weight, moisture, stdMoisture = 13.5, barley = FALSE,
  ...)
}
\arguments{
\item{weight}{[value]}

\item{moisture}{[value]}

\item{stdMoisture}{numeric. moisture percentage to standardize grain yield.}

\item{barley}{logical. Is it barley?}
}
\value{
[value]
}
\description{
function to correct weight based on percent moisture. Standard moisture is 13.5 for wheat and 14.5 for barley
}
\details{
[fill in details here]
}
\examples{
none
}
