% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_keywords.R
\name{define_keywords}
\alias{define_keywords}
\title{Modify Keywords Used In Outputs}
\usage{
define_keywords()
}
\description{
As an alternative to \code{\link{use_custom_lang}}, this allows temporarily
modifying the keywords used in the outputs.
}
\details{
A window will pop-up, allowing the modification of the \emph{custom} column.
The changes will be active as long as the package is loaded.
}
\keyword{utilities}
