% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfSummary_simple.R
\name{dfSummary_simple}
\alias{dfSummary_simple}
\title{dfSummary_simple}
\usage{
dfSummary_simple(bdd, nom_var, nom_croisement = NULL, poids = NULL,
  filtres = NULL, header_perso = TRUE)
}
\arguments{
\item{bdd}{Completer}

\item{nom_var}{Completer}

\item{nom_croisement}{Completer}

\item{poids}{Completer}

\item{filtres}{Completer}

\item{header_perso}{Completer}
}
\description{
dfSummary_simple
}
\author{
antuki
}
