% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfSummary.R
\name{format_number}
\alias{format_number}
\title{format_number}
\usage{
format_number(x, round.digits, ...)
}
\arguments{
\item{x}{A numerical value to be formatted.}

\item{round.digits}{Numerical. Number of decimals to show. Used to define 
both \code{digits} and \code{nsmall} when calling \code{\link{format}}.}

\item{...}{Any other formatting instruction that is compatible with 
\code{\link{format}}.}
}
\description{
Used internally (not exported) to apply all relevant formatting. It is 
documented here only because it can be used when setting the 
\code{dfSummary.custom.1} and  \code{dfSummary.custom.1} options.
}
