% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatPrintSummary2.R
\name{formatPrintSummary2}
\alias{formatPrintSummary2}
\title{formatPrintSummary function}
\usage{
formatPrintSummary2(smry, infoCols = "Line", tex = TRUE,
  sortBy = NULL, loHi = FALSE, decimals = c(3, 4))
}
\arguments{
\item{smry}{summary output from oneYearOneLocSummary, oneYearOverLocSummary, or multiYearSummary}

\item{infoCols}{list of col names to be included in printed list as is}

\item{tex}{should a latex table be printed to stdout?}
}
\value{
[value]
}
\description{
function to sort summaries.
}
\details{
[fill in details here]
}
\examples{
# none
}
