% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatTime.R
\name{formatTime}
\alias{formatTime}
\title{formatTime function}
\usage{
formatTime(x, year)
}
\arguments{
\item{x}{[value]}

\item{year}{[value]}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
# none
}
