% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatXtabColNames.R
\name{formatXtabColNames}
\alias{formatXtabColNames}
\title{formatXtabColNames function}
\usage{
formatXtabColNames(x, unitSep = "\\\\|")
}
\arguments{
\item{x}{character vector of format 'Trait|unit (#trials)'}

\item{unitSep}{unit sep (must have escape characters if neede, i.e. '|' would be "\\|")}
}
\value{
[value]
}
\description{
function to format data.frame names from 'Trait|unit (#trials)' to multi lines for printing in xtable
}
\details{
[fill in details here]
}
\examples{
# none
}
