% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitConversions.R
\name{gltolbbu}
\alias{gltolbbu}
\title{gltolbbu function}
\usage{
gltolbbu(gl)
}
\arguments{
\item{gl}{[value]}
}
\value{
[value]
}
\description{
function to convert testweigth from grams / liter to pounds / bushel
}
\details{
[fill in details here]
}
\examples{
# none
}
