% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitConversions.R
\name{gramsToYield}
\alias{gramsToYield}
\title{lbsToYield function}
\usage{
gramsToYield(g, moisture, sqft = NULL, sqm = NULL, metric = FALSE,
  barley = FALSE, ...)
}
\arguments{
\item{g}{numeric. Grain mass in grams.}

\item{moisture}{numeric. grain moisture as a percent, 0 < moisture < 100}

\item{sqft}{numeric. plot area in square feet.}

\item{sqm}{numeric. plot area in square meters.}

\item{metric}{logical. should metric units be returned?}

\item{barley}{logical. Is it barley?}
}
\value{
[value]
}
\description{
function to convert pounds to grain yield using standard moisture
}
\details{
[fill in details here]
}
\examples{
none
}
