% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitConversions.R
\name{inchToCm}
\alias{inchToCm}
\title{inchToCm function}
\usage{
inchToCm(inches)
}
\arguments{
\item{inches}{[value]}
}
\value{
[value]
}
\description{
function to convert inches to centimeters.
}
\details{
[fill in details here]
}
\examples{
none
}
