% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jeanetteSummary.R
\name{jeanetteSummary}
\alias{jeanetteSummary}
\title{jeanetteSummary function}
\usage{
jeanetteSummary(path, checks, nWide = 5, inEnv = FALSE, keep = NULL,
  ignore = NULL, printMeans = TRUE, byLoc = "YLD_BUPA",
  shorthand = TRUE, sdigits = NULL, ...)
}
\arguments{
\item{path}{file path}

\item{checks}{checks to be printed, only used if inEnv = TRUE}

\item{nWide}{number of tables wide for inEnv table printing, default is 5}

\item{inEnv}{logical. should means within environment be printed? default is = FALSE}

\item{keep}{traits to keep, default is NULL}

\item{ignore}{traits to ignore, default is NULL. Note hasnt been implemented yet}

\item{printMeans}{logical. Should he means be printeed? Default is TRUE. Specify a file path to 'con' to write to a file, else tex tables will be printed to stdout.}

\item{byLoc}{trait for which to prodce tables by location, default is "YLD".}

\item{shorthand}{logical. Should shorthand for traits be used? Makes tables tighter. default is TRUE,}

\item{sdigits}{numeric. vector of number of digits for single environment traits. for multi environment traits, use 'digits' argument whcih is passed to formatPrintSummary. default is null, and no rounding will be accomplished}

\item{nlineHeader}{number of lines in the header}
}
\value{
[value]
}
\description{
function to read Jeanette Lyerly summaries and produce summaries over locations.
}
\details{
[fill in details here]
}
\examples{
# none
}
