% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitConversions.R
\name{kghaToBuac}
\alias{kghaToBuac}
\title{buacToLbkg function}
\usage{
kghaToBuac(kgha, barley = FALSE)
}
\arguments{
\item{barley}{logical. Is it barley?}

\item{bu}{numeric. Bushels.}
}
\value{
[value]
}
\description{
function to convert kilograms per hectare to bushels per acre based on standard testweight (60 for wheat and 48 for barley)
}
\details{
[fill in details here]
}
\examples{
# none
}
