% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitConversions.R
\name{kgToBu}
\alias{kgToBu}
\title{kgToBu function}
\usage{
kgToBu(kg, barley = FALSE)
}
\arguments{
\item{kg}{numeric. Bushels.}

\item{barley}{logical. Is it barley?}
}
\value{
[value]
}
\description{
function to convert bushels to kilograms based on standard testweight (60 for wheat and 48 for barley)
}
\details{
[fill in details here]
}
\examples{
# none
}
