% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label.R
\name{label<-}
\alias{label<-}
\title{Set Variable or Data Frame Label}
\usage{
label(x) <- value
}
\arguments{
\item{x}{The data frame or variable to be labelled.}

\item{value}{String to be used as label.}
}
\description{
Defines a data frame or variable label by using the \code{label} attribute.
}
\seealso{
\code{\link{label}} (\pkg{base}), \code{\link[rapportools]{label}} (\pkg{rapportools}),
  \code{\link[Hmisc]{label}} (\pkg{Hmisc})
}
\author{
Dominic Comtois, \email{dominic.comtois@gmail.com}
}
