% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitConversions.R
\name{lbbutogl}
\alias{lbbutogl}
\title{lbbutogl function}
\usage{
lbbutogl(lbbu)
}
\arguments{
\item{lbbu}{[value]}
}
\value{
[value]
}
\description{
function to convert testweigth from pounds / bushel to grams / liter
}
\details{
[fill in details here]
}
\examples{
none
}
