% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitConversions.R
\name{lbsToYield}
\alias{lbsToYield}
\title{lbsToYield function}
\usage{
lbsToYield(lbs, moisture, sqft, barley = FALSE, ...)
}
\arguments{
\item{moisture}{numeric. grain moisture as a percent, 0 < moisture < 100}

\item{sqft}{numeric. plot area in square feet.}

\item{barley}{logical. Is it barley?}

\item{weight}{numeric. grain weight in pounds.}
}
\value{
[value]
}
\description{
function to convert pounds to grain yield using standard moisture
}
\details{
[fill in details here]
}
\examples{
# none
}
