% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make4colScale.R
\name{make4colScale}
\alias{make4colScale}
\title{make4colScale function}
\usage{
make4colScale(scl3col)
}
\arguments{
\item{scl3col}{[value]}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
# none
}
