% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make9colDJ.R
\name{make9colDJ}
\alias{make9colDJ}
\title{make9colDJ function}
\usage{
make9colDJ(dj3col)
}
\arguments{
\item{dj3col}{[value]}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
