% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePurdy.R
\name{makePurdy}
\alias{makePurdy}
\title{makePurdy function}
\usage{
makePurdy(female, male, includeSpace = TRUE)
}
\arguments{
\item{female}{pedigree with symbol between female and male pedigrees}

\item{male}{character vector of variables to add dummy to}
}
\value{
data.frame with added dummy plots to make square.
}
\description{
function to make purdy pedigrees from female and male pedigres or line names.
}
\details{
[fill in details here]
}
\examples{
makePurdy("A/B", "C")
makePurdy("A/B", "C", includeSpace = FALSE)
makePurdy("A/B", "C/D//E", includeSpace = FALSE)
}
