% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSquare.R
\name{makeSquare}
\alias{makeSquare}
\title{makeSquare function}
\usage{
makeSquare(
  dF,
  dvars = NULL,
  range = "range",
  pass = "pass",
  dummy = "dummy",
  imputeFactors = TRUE,
  by = NULL
)
}
\arguments{
\item{dF}{data.frame of field trial,}

\item{dvars}{character vector of variables to add dummy to}

\item{range}{character. data.frame variable name to indicate row. Default is 'range'}

\item{pass}{character. data.frame variable name to indicate column. Default is 'pass'}

\item{dummy}{character. Stand in value for added plots to make square. Default is 'dummy'}

\item{by}{character. dF column name by which to apply square, usually physically  separate locations, blocks, etc.}
}
\value{
data.frame with added dummy plots to make square.
}
\description{
function to add dummy plots to a field trial to make it square. Useful for spatial analysis.
}
\details{
[fill in details here]
}
\examples{
# none
}
