% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTraitNameUnit.R
\name{makeTraitNameUnit}
\alias{makeTraitNameUnit}
\title{cleanTraitNames function}
\usage{
makeTraitNameUnit(traitNames, units, unitSep = "|")
}
\arguments{
\item{traitNames}{[value] trait names without units}

\item{units}{[value]. units for triats, all traits with units of an empty string will not have units assigned}

\item{unitSep}{[value]. Pattern that separates triat name from units or ontology}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
# none
}
