% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchTraits.R
\name{matchTraits}
\alias{matchTraits}
\title{matchTraits function}
\usage{
matchTraits(traitNames, traits, warnMultiple = TRUE)
}
\arguments{
\item{traitNames}{character vector of trait names, typically as column names in a summary data.frame, these would often have a pipe, "|", and a unit, and possibly the number of environments in parentheses after the trait name which would prevent direct matching with more simple trait names, Example, "GrainYield|lb/ac (12)"}

\item{traits}{character vector of traits, typically these would not have units or the number of environments, e.g. "Grain Yield"}
}
\value{
[value]
}
\description{
function to match trait names with or without units, numbers of environments, etc. Spaces are ignored, but error can result if one trait has a name that is nested within another, e.g 'Plant' and 'PlantHeight'.
}
\details{
[fill in details here]
}
\examples{
# none
}
