% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeFieldBookData.R
\name{mergeFieldBookData}
\alias{mergeFieldBookData}
\title{mergeFieldData function}
\usage{
mergeFieldBookData(fb, dj = NULL, scl = NULL, paperfb = NULL,
  trialDesigns = NULL, trialPassRange = NULL, sqft = 45,
  testName = NULL, year = NULL, inclPed = TRUE, ...)
}
\arguments{
\item{fb}{[value]}

\item{dj}{[value]}

\item{scl}{[value]}

\item{sqft}{[value]. Default is 45}

\item{testName}{[value]. Default is NULL}

\item{year}{[value]. Default is NULL}
}
\value{
[value]
}
\description{
function to merge fieldbook data collected with the fieldbook application.
}
\details{
[fill in details here]
}
\examples{
none
}
