% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeFieldData.R
\name{mergeFieldData}
\alias{mergeFieldData}
\title{mergeFieldData function}
\usage{
mergeFieldData(fb, dj, scl, sqft = 45, testName = NULL, year = NULL,
  ...)
}
\arguments{
\item{fb}{[value]}

\item{dj}{[value]}

\item{scl}{[value]}

\item{sqft}{[value]. Default is 45}

\item{testName}{[value]. Default is NULL}

\item{year}{[value]. Default is NULL}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
