% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeHarvM.R
\name{mergeHarvM}
\alias{mergeHarvM}
\title{mergeHarvM function}
\usage{
mergeHarvM(hmL, sclL, djL)
}
\arguments{
\item{hmL}{[value]}

\item{sclL}{[value]}

\item{djL}{[value]}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
