% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeNotes.R
\name{mergeNotes}
\alias{mergeNotes}
\title{mergreNotes function}
\usage{
mergeNotes(BLUtab, notes)
}
\arguments{
\item{BLUtab}{summary of line effects with stats at bottom, output from makeBLUtab}

\item{compiled}{notes across plots for each line, output from compileNotes}
}
\value{
summary table with notes added, and no stats lost.
}
\description{
function to merge compiled notes with summary tables without losing stats.
}
\details{
[fill in details here]
}
\examples{
# none
}
