% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeSummary.R
\name{mergeSummary}
\alias{mergeSummary}
\title{readRegionalExcel function}
\usage{
mergeSummary(x, y, ...)
}
\arguments{
\item{x}{first summary to be merged}

\item{y}{second summary to be merged}
}
\value{
[value]
}
\description{
function to read Wade's crazy excel table summaries.
}
\details{
[fill in details here]
}
\examples{
none
}
