% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiYearSummary.R
\name{multiYearSummary}
\alias{multiYearSummary}
\title{multiYearSummary function}
\usage{
multiYearSummary(dF, traits, locs = NULL, sortby = NULL,
  allowDupEnt = TRUE, blockName = "Bloc", ...)
}
\arguments{
\item{dF}{[value]}

\item{traits}{[value]}

\item{locs}{[value]. Default is NULL}

\item{sortby}{[value]. Default is NULL}

\item{allowDupEnt}{[value]. Default is TRUE}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
# none
}
