% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numAwnToChar.R
\name{numAwnToChar}
\alias{numAwnToChar}
\title{numAwnToChar function}
\usage{
numAwnToChar(x)
}
\arguments{
\item{x}{vector of awn scores, either numeric, character or mixed}
}
\value{
[value]
}
\description{
function to select an awn type from multiple observations. Intended to be used from a call from cleanAwns(). TA/AL conflicts are assumed to be TA, while TA/A  are assumed to be awned. A/AL conflicts will be printed as A,AL
}
\details{
[fill in details here]
}
\examples{
# none
}
