% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/numericSummary.R
\name{numericSummary}
\alias{numericSummary}
\title{Generate numeric data summaries}
\usage{
numericSummary(vec, na.rm = TRUE)
}
\arguments{
\item{vec}{Numeric vector of data to generate data summaries}

\item{na.rm}{Should missing values be removed?}
}
\value{
This function will return a named vector containing:
\describe{
 \item{Min}{The minimum value of the data}
 \item{Mean}{The mean of the data}
 \item{Variance}{The variance of the data}
 \item{Max}{The maximum value of the data}
 \item{Nas}{The number of missing values}
}
}
\description{
Generate numeric data summaries
}
\examples{
numericSummary(airquality$Ozone)
}
\author{
Aimee Gott
}

