% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oneYearOneLocSummary.R
\name{oneYearOneLocSummary}
\alias{oneYearOneLocSummary}
\title{oneYearOneLocSummary function}
\usage{
oneYearOneLocSummary(dF, traits, sortHiLo = NULL, sortLoHi = NULL,
  allowDupEnt = TRUE, unitSep = "|", verbose = FALSE, fixed = NULL,
  random = NULL, printFit = FALSE, ...)
}
\arguments{
\item{dF}{data.frame of trial data. Variables "Trial", "Line", "Entry", "Block", "plot_name" must be included. plot_name is must be of the form Test_Year_Location.}

\item{traits}{character vector of traits to analyze.}

\item{sortHiLo}{Trait by which to sort summary high to low. Leave NULL to sort by entry. Default is NULL}

\item{sortLoHi}{Trait by which to sort summary low to high. Leave NULL to sort by entry. Default is NULL}

\item{allowDupEnt}{logical. Should duplicated entries be allowed? Useful for the same line from different seed sources. default is TRUE}

\item{unitSep}{character. separator for units. Typically this is either a space, parenthesis, or in line with T3, a pipe. default is "|"}

\item{verbose}{logical. Default is TRUE}

\item{fixed}{[value]. Default is NULL}

\item{random}{[value]. Default is NULL}

\item{printFit}{logical. Should the model fit be printed to stdout? Default is FALSE}
}
\value{
[value]
}
\description{
function to analyze trial data and return a trial summary with lines treated as fixed (BLUE) and random (BLUP). LSD and CV is returned with the BLUEs and a broad-sense heritability and the error standard deviation are returned with the BLUPs.
}
\details{
[fill in details here]
}
\examples{
# none
}
