% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFieldBook.R
\name{readFieldBook}
\alias{readFieldBook}
\title{readFieldBook function}
\usage{
readFieldBook(path, year, testName = NULL, sheets = NULL,
  colOrder = NULL, scoreTraits = NULL)
}
\arguments{
\item{path}{[value]}

\item{year}{[value]}

\item{testName}{[value]. Default is NULL}

\item{sheets}{[value]. Default is NULL}

\item{colOrder}{[value]. Default is NULL}

\item{scoreTraits}{[value]. Default is NULL}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
