% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFieldBookDB.R
\name{readFieldBookDB}
\alias{readFieldBookDB}
\title{readFieldBookDB function}
\usage{
readFieldBookDB(path, ontology = NULL, printTraitTable = FALSE,
  exportNA = FALSE, printDup = FALSE, exportFullFB = FALSE)
}
\arguments{
\item{path}{character. Path to file or directory containing multiple database files exported from Fieldbook app.}

\item{ontology}{data.frame. data.frame with two variables, 'pattern' as the regrular expression to match with each each trait, paired to an 'ontology', which should have a pipe and an ontology number (eg. 'my trait|CO123:456789')}

\item{printTraitTable}{logical, should the table of traits be printed? default is FALSE,}

\item{exportNA}{logical. should records with missing information be printed? default is FALSE}

\item{printDup}{logical. should duplicate records be printed? default is FALSE}

\item{exportFullFB}{logical. Should the entire fieldbook be exported (i.e. no duplicate or missing data records removed)? default is  FALSE}
}
\value{
data.frame of processed and sorted field book records
}
\description{
function to read database files exported by the Fieldbook Application
}
\details{
[fill in details here]
}
\examples{
none
}
