% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readRegionalExcel.R
\name{readRegionalExcel}
\alias{readRegionalExcel}
\title{readRegionalExcel function}
\usage{
readRegionalExcel(path, nlineHeader = 2, ...)
}
\arguments{
\item{path}{file path}

\item{nlineHeader}{number of lines in the header}
}
\value{
[value]
}
\description{
function to read Wade's crazy excel table summaries.
}
\details{
[fill in details here]
}
\examples{
# none
}
