% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readWadeExcel.R
\name{readWadeExcel}
\alias{readWadeExcel}
\title{readWadeExcel function}
\usage{
readWadeExcel(
  path,
  startLine = NULL,
  stopLine = NULL,
  keepLSD = FALSE,
  tooLong = 33,
  ...
)
}
\arguments{
\item{path}{file path}

\item{startLine}{optional start line}

\item{stopLine}{optional stop line}

\item{keepLSD}{should the plus Minus LSD separation from mean be kept?}

\item{tooLong}{parameter to determine if the first col name is just a title}
}
\value{
[value]
}
\description{
function to read Wade's crazy excel table summaries.
}
\details{
[fill in details here]
}
\examples{
# none
}
