% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recodeLoc.R
\name{recodeLoc}
\alias{recodeLoc}
\title{recodeLoc function}
\usage{
recodeLoc(loc, locCode = NULL, locPattern = NULL)
}
\arguments{
\item{loc}{[value]}

\item{locCode}{[value]. Default is NULL}

\item{locPattern}{[value]. Default is NULL}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
