% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rightpedsym.R
\name{rightpedsym}
\alias{rightpedsym}
\title{makeSquare function}
\usage{
rightpedsym(ped, pattern = "\\\\s*/x/\\\\s*", includeSpace = TRUE)
}
\arguments{
\item{ped}{pedigree with symbol between female and male pedigrees}

\item{dvars}{character vector of variables to add dummy to}

\item{range}{character. data.frame variable name to indicate row. Default is 'range'}

\item{pass}{character. data.frame variable name to indicate column. Default is 'pass'}

\item{dummy}{character. Stand in value for added plots to make square. Default is 'dummy'}

\item{by}{character. dF column name by which to apply square, usually physically  separate locations, blocks, etc.}
}
\value{
data.frame with added dummy plots to make square.
}
\description{
function to assign purdy pedigrees symbols correctly when parents are expressed as pedigrees (e.g. A/B crossed to C gives A/B//C). Needs the user to put a recognizable symbol between female ad male pedigrees, such as /x/
}
\details{
[fill in details here]
}
\examples{
rightpedsym("A/B /x/ C")
rightpedsym("A/B /x/ C", includeSpace=FALSE)
rightpedsym("A x B/C", pattern = "x", includeSpace=FALSE)
rightpedsym("A x B/C", pattern = " x ", includeSpace=FALSE)
}
