% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roundSummary.R
\name{roundSummary}
\alias{roundSummary}
\title{roundSummary function}
\usage{
roundSummary(smry, dig1 = "Test|Yield")
}
\arguments{
\item{smry}{summary with mean, cv and lsd at end. Need plus minus to know which columns to round}

\item{dig1}{use 1 digit for these traits that match this regex pattern}
}
\value{
[value]
}
\description{
function to round summaries. Only works if plus minus present
}
\details{
[fill in details here]
}
\examples{
none
}
