% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitConversions.R
\name{scoreToAwns}
\alias{scoreToAwns}
\title{scoreToAwns function}
\usage{
scoreToAwns(awnScore)
}
\arguments{
\item{awnScore}{[value]}
}
\value{
[value]
}
\description{
function to convert awn scores (1, 2, 3) to awns (A, TA, AL)
}
\details{
[fill in details here]
}
\examples{
# none
}
