% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectAwn.R
\name{selectAwn}
\alias{selectAwn}
\title{selectAwn function}
\usage{
selectAwn(x)
}
\arguments{
\item{x}{list of multiple awn observations per element e.g.(list(c("A", "TA"), c("A", "A", "A")))}
}
\value{
[value]
}
\description{
function to select an awn type from multiple observations. Intended to be used from a call from cleanAwns(). TA/AL conflicts are assumed to be TA, while TA/A  are assumed to be awned. A/AL conflicts will be printed as A,AL
}
\details{
[fill in details here]
}
\examples{
# none
}
