% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortSummary.R
\name{sortSummary}
\alias{sortSummary}
\title{sortSummary function}
\usage{
sortSummary(smry, trait, decreasing = TRUE, ...)
}
\arguments{
\item{smry}{summary output from oneYearOneLocSummary, oneYearOverLocSummary, or multiYearSummary}

\item{decreasing}{should sort be done by decreasing values?}

\item{traits}{list of traits to provide LSD from mean to}
}
\value{
[value]
}
\description{
function to sort summaries.
}
\details{
[fill in details here]
}
\examples{
# none
}
