% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitConversions.R
\name{sqftToAcre}
\alias{sqftToAcre}
\title{sqftToAcre function}
\usage{
sqftToAcre(sqft)
}
\arguments{
\item{sqft}{[value]}
}
\value{
[value]
}
\description{
function to convert square feet to acres
}
\details{
[fill in details here]
}
\examples{
none
}
