% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitConversions.R
\name{sqftToHectare}
\alias{sqftToHectare}
\title{sqftToHectare function}
\usage{
sqftToHectare(sqft)
}
\arguments{
\item{sqft}{[value]}
}
\value{
[value]
}
\description{
function to convert square feet to hectares
}
\details{
[fill in details here]
}
\examples{
# none
}
