% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statSummary.R
\name{statSummary}
\alias{statSummary}
\title{Statistical Summaries}
\usage{
statSummary(x, na.rm = TRUE)
}
\arguments{
\item{x}{numeric vector}

\item{na.rm}{logical. Remove NAs}
}
\value{
A named numeric vector
}
\description{
Statistical Summaries
}
\examples{
statSummary(iris$Sepal.Length)
}

