% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strPad.R
\name{strPad}
\alias{strPad}
\title{strPad function}
\usage{
strPad(x, type = "l", pad = 0)
}
\arguments{
\item{x}{characer vector to be padded}

\item{type}{either r, l or c for left , right or center.}

\item{pad}{should additional pad be added?}
}
\value{
[value]
}
\description{
function to pad strings.
}
\details{
[fill in details here]
}
\examples{
# none
}
