% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summaryTools-package.R
\docType{package}
\name{summaryTools}
\alias{summaryTools}
\alias{summaryTools-package}
\title{summaryTools: A package for generating standard numeric summaries}
\description{
This package will allow you to generate a standard set of summary statistics
for vectors of data and has been designed for the purpose of demonstrating
package building methods and structure.
}

