% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabagisme.R
\docType{data}
\name{tabagisme}
\alias{tabagisme}
\title{Usage du Tabac et état de Santé (Données simulées)}
\format{Un data frame de 1000 rangées et 9 colonnes}
\usage{
data(tabagisme)
}
\description{
Jeu de données simulées de 1000 sujets, avec les
colonnes suivantes:
\itemize{
  \item sexe Variable catégorielle (facteur), 2 niveaux:
    \dQuote{F} et \dQuote{M}. Environ 500 chacun.
  \item age Numérique.
  \item age.gr Age regroupé en facteur (4 niveaux).
  \item IMC Indice de masse corporelle (numérique).
  \item fumeur Variable catégorielle à 2 niveaux
    (\dQuote{Oui} / \dQuote{Non}).
  \item cigs.par.jour Nombre de cigarettes fumées par jour
    (numérique).
  \item malade Variable catégorielle à 2 niveaux
    (\dQuote{Oui} / \dQuote{Non}).
  \item maladie Champs texte.
  \item ponderation Poids échantillonal (numérique).
}
}
\details{
Note sur la simulation des données: la probabilité pour
un sujet de tomber dans la catégorie \dQuote{malade} est
basée sur une fonction arbitraire faisant intervenir l'âge,
l'IMC et le nombre de cigarettes fumées par jour.

A copy of this dataset is \strong{available in English} under the name
\dQuote{tobacco}.
}
\keyword{datasets}
