% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trialDesignToDataFrame.R
\name{trialDesignToDataFrame}
\alias{trialDesignToDataFrame}
\title{trialDesignToDataFrame function}
\usage{
trialDesignToDataFrame(tr, inclPed = TRUE)
}
\arguments{
\item{tr}{object of class trialDesign.}
}
\value{
data.frame with plot_name, trialName, Year, Location, Line, Entry, and Block.
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
