% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label.R
\name{unlabel}
\alias{unlabel}
\title{Clear Variable or Data Frame Label(s)}
\usage{
unlabel(x)
}
\arguments{
\item{x}{An R object to remove labels from.}
}
\description{
Returns the object with all labels removed. Both the \dQuote{label} attribute
and \pkg{Hmisc}'s \dQuote{labelled} class are removed.
}
\details{
Can be more practical than using \code{label(x) <- NA}, if we want to keep
the label but not let it interfere with one particular task.
}
\seealso{
\code{link{label}}, \code{\link[rapportools]{label}},
  \code{\link[Hmisc]{label}}
}
\author{
Dominic Comtois, \email{dominic.comtois@gmail.com},
}
