% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_custom_lang.R
\name{use_custom_lang}
\alias{use_custom_lang}
\title{Import and use a custom language}
\usage{
use_custom_lang(file)
}
\arguments{
\item{file}{Character. The path to the translations file.}
}
\description{
If your language is not available or if you wish to customize the outputs'
language to suit your preference, you can set up a translations file (see
details) and import it with this function.
}
\details{
To build the translations file, download
\href{https://raw.githubusercontent.com/dcomtois/summarytools/master/translations/language_template.csv}{this template},
or copy the \emph{language_template.csv} file located in the installed 
package’s \emph{includes} directory and fill out the \sQuote{custom} column
using a text editor, leaving column titles unchanged. The file must also
retain its \emph{UTF-8} encoding.
}
\keyword{utilities}
