% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whichTrials.R
\name{whichTrials}
\alias{whichTrials}
\title{whichTrials function}
\usage{
whichTrials(dF, trait)
}
\arguments{
\item{dF}{[value]}

\item{trait}{[value]}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
# none
}
