% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.fbData.R
\name{write.fbData}
\alias{write.fbData}
\title{write.fbData function}
\usage{
write.fbData(db, path = NULL)
}
\arguments{
\item{db}{data.frame out put from dataToFb function}

\item{path}{optional chaacter vector of path to where he file should be written}
}
\value{
[value]
}
\description{
function to convert a data.frame of data o fieldook database file. Usual used to get data from colaboators, etc, anyting that was recored on a spreadshet.
}
\details{
[fill in details here]
}
\examples{
# none
}
