% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.summary.R
\name{write.summary}
\alias{write.summary}
\title{write.summary function}
\usage{
write.summary(locSum, testName, resultDir, digits = 2, addLSD = FALSE,
  ...)
}
\arguments{
\item{locSum}{[value]}

\item{testName}{[value]}

\item{resultDir}{[value]}

\item{digits}{[value]. Default is 2}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
# none
}
