% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeFieldBook.R
\name{writeFieldBook}
\alias{writeFieldBook}
\title{writeFieldBook function}
\usage{
writeFieldBook(testDf, traits, keepCols = NULL, unitSep = "|")
}
\arguments{
\item{testDf}{[value]}

\item{traits}{[value]}

\item{keepCols}{[value]. Default is NULL}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
