% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xtable2.R
\name{xtable2}
\alias{xtable2}
\title{xtable2 function}
\usage{
xtable2(tab, aln = NULL, pad = 0, digits = 2, caption = NULL,
  label = NULL, size = "footnotesize", footNotes = NULL,
  include.rownames = TRUE, formatColNames = TRUE, statsLine = TRUE,
  specialChars = c("#"), ...)
}
\arguments{
\item{tab}{table to be printed into a latex table}

\item{aln}{character of length 1, indicating alignment, l, r or c.}

\item{pad}{extra pad to be added}

\item{digits}{vector of length ncol(tab) indicating the number of digits. If columns are not nueric, you still have to add some value for each column, which will be ignored.}

\item{caption}{latex latex caption. If you complicated you might want to skip this and edit the output tex table}

\item{label}{latex latex label name}

\item{size}{latex textsize. do not add slashes before, they will be appended automatically.}

\item{formatColNames}{will reformat column names of the type 'Trait|unit (#trials)' to multiple lines, 'Trait \\ unit \\ #trials' for better printing}

\item{statsLine}{should a line be printed before mean, CV, LSD?}

\item{ignore.rownames}{should rownames of the table be ignored? default is TRUE}
}
\value{
[value]
}
\description{
function to print latex table with padding for easier reading / manual manipulation thna xtable (xtable may have this feature?)
}
\details{
[fill in details here]
}
\examples{
# none
}
