% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addOntology.R
\name{addOntology}
\alias{addOntology}
\title{addOntology function}
\usage{
addOntology(testData, ontology, exactMatch = FALSE, returnDF = FALSE)
}
\arguments{
\item{testData}{data.frame of test data with traits in columns.}

\item{ontology}{data.frame with two columns, first must be the trait pattern to match and the second is the ontology}

\item{exactMatch}{logical. match the trait pattern exactly. Default is FALSE}

\item{returnDF}{logical. should a dataframe be returned? Default is FALSE}
}
\value{
[value]
}
\description{
function to change trait names to add ontology numbers.
}
\details{
[fill in details here]
}
\examples{
# none
}
