% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanTraitNames.R
\name{cleanTraitNames}
\alias{cleanTraitNames}
\title{cleanTraitNames function}
\usage{
cleanTraitNames(traitNames, splitPattern = "\\n|\\\\|")
}
\arguments{
\item{traitNames}{[value]}

\item{splitPattern}{[value]. Pattern that separates triat name from units or ontology}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
# none
}
