% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleartmp.R
\name{cleartmp}
\alias{cleartmp}
\title{Delete Temporary Html Files}
\usage{
cleartmp(all = FALSE, silent = FALSE)
}
\arguments{
\item{all}{Logical. When \code{TRUE}, all temporary summarytools are deleted.
When \code{FALSE} (default), only the latest is.}

\item{silent}{Hide confirmation messages (\code{FALSE} by default).}
}
\description{
Delete temporary files created when using generic print method with 
\code{method='browser'} or \code{method='viewer'}, or when calling
\code{view()} function.
}
\details{
All temporary files are deleted automatically when R session is ended. This
function is thus an overkill in most circumstances.
}
\author{
Dominic Comtois, \email{dominic.comtois@gmail.com>}
}
\keyword{IO}
