% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleartmp.R
\name{cleartmp}
\alias{cleartmp}
\title{Delete Temporary Html Files}
\usage{
cleartmp(all = TRUE, silent = FALSE, verbose = FALSE)
}
\arguments{
\item{all}{Logical. When \code{TRUE} (default), all temporary summarytools 
files are deleted. When \code{FALSE}, only the latest file is.}

\item{silent}{Logical. Hide confirmation messages (\code{FALSE} by default).}

\item{verbose}{Logical. Display a message for every file that is deleted.
\code{FALSE} by default.}
}
\description{
Delete temporary files created when using generic print method with 
\code{method='browser'} or \code{method='viewer'}, or when calling
\code{view()} function.
}
\note{
Given that all temporary files are deleted automatically when an R
session is ended, this function is an overkill in most circumstances. It 
could however be useful in server-type setups.
}
\author{
Dominic Comtois, \email{dominic.comtois@gmail.com>}
}
\keyword{IO}
