% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctable.R
\name{ctable}
\alias{ctable}
\title{Cross-Tabulation}
\usage{
ctable(x, y, prop = "r", useNA = "ifany", totals = TRUE,
  style = "simple", round.digits = 1, justify = "right",
  omit.headings = FALSE, plain.ascii = TRUE, split.tables = Inf,
  dnn = c(substitute(x), substitute(y)), ...)
}
\arguments{
\item{x}{First categorical variable - values will appear as row names.}

\item{y}{Second categorical variable - values will appear in as column names.}

\item{prop}{Proportions to display;  \dQuote{r} for \emph{rows} (default),
\dQuote{c} for \emph{columns}, \dQuote{t} for \emph{total}, or \dQuote{n} for
\emph{none}.}

\item{useNA}{Argument passed on to \code{\link[base]{table}}; One of \dQuote{ifany}
(default), \dQuote{no}, or \dQuote{always}.}

\item{totals}{Logical. Should row and column totals be displayed? Defaults to \code{TRUE}.}

\item{style}{Style to be used by \code{\link[pander]{pander}} when rendering
output table; One of \dQuote{simple} (default), \dQuote{grid}, or \dQuote{rmarkdown} 
(see details section about the latter).}

\item{round.digits}{Number of significant digits to display. Defaults to
\code{1}.
\dQuote{rmarkdown}.}

\item{justify}{String indicating alignment of columns; one of \dQuote{l} (left)
\dQuote{c} (center), or \dQuote{r} (right). Defaults to \dQuote{r}.}

\item{omit.headings}{Logical. Set to \code{TRUE} to omit headings.}

\item{plain.ascii}{Logical. \code{\link[pander]{pander}} argument; when
\code{TRUE}, no markup characters will be used (useful when printing
to console). Defaults to \code{TRUE} when \code{style} is \dQuote{simple},
and \code{FALSE} otherwise.}

\item{split.tables}{Pander argument that specifies how many characters wide a
table can be. \code{Inf} by default.}

\item{dnn}{Names to be used in output table. Vector of two strings; By default,
the character values for arguments x and y are used.}

\item{\dots}{Additional arguments passed to \code{\link[pander]{pander}}.}
}
\value{
A frequency table of class \code{matrix} with added attributes used
  by \link{print} method.
}
\description{
Cross-tabulation for a pair of categorical variables (or factors) with either row,
column, or total proportions, as well as marginal sums.
}
\details{
Rmarkdown does not, to this day, support multi-header tables. Therefore,
  until such support is available, the recommended way to display cross-tables
  in .Rmd documents is to use `method=render` with the `print()` generic function
  or with the `view()` function. See package vignettes for examples.
}
\examples{
data("tobacco")
ctable(tobacco$gender, tobacco$smoker)

}
\seealso{
\code{\link[base]{table}}, \code{\link[stats]{xtabs}}
}
\author{
Dominic Comtois, \email{dominic.comtois@gmail.com}
}
\keyword{category}
\keyword{classes}
