% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctable.R
\name{ctable}
\alias{ctable}
\title{Cross-Tabulations}
\usage{
ctable(x, y, prop = "t", totals = TRUE, round.digits = 1,
  useNA = "ifany", style = "simple", plain.ascii = TRUE,
  justify = "right", dnn = c(substitute(x), substitute(y)), ...)
}
\arguments{
\item{x}{First categorical variable - its values will appear as row names.}

\item{y}{Second categorical variable - its values will appear in as column names.}

\item{prop}{What proportions to display; \dQuote{t} for \emph{total} (default),
\dQuote{r} for \emph{rows}, \dQuote{c} for \emph{columns} or \dQuote{n} for
\emph{None}.}

\item{totals}{Should row and column totals be displayed? Defaults to \code{TRUE}.}

\item{round.digits}{Number of significant digits to display. Defaults to
\code{1}.}

\item{useNA}{Argument used by \code{\link[base]{table}}; One of \dQuote{ifany}
(default), \dQuote{no}, or \dQuote{always}.}

\item{style}{Style to be used by \code{\link[pander]{pander}} when rendering
output table; One of \dQuote{simple} (default), \dQuote{grid} or
\dQuote{rmarkdown}.}

\item{plain.ascii}{Logical \code{\link[pander]{pander}} argument. When
\code{TRUE}, no markup characters will be used (useful when printing
to console). Defaults to \code{TRUE} when \code{style} is \dQuote{simple},
and \code{FALSE} otherwise.}

\item{justify}{String indicating alignment of columns; one of \dQuote{left}
(or \dQuote{l}), \dQuote{center} (or \dQuote{c}), or \dQuote{right}
(or \dQuote{r}). Defaults to \dQuote{right}.}

\item{dnn}{Names to be used in output table. Vector of two strings; By default,
the character values for arguments x and y are used.}

\item{\dots}{Additional arguments passed to \code{\link[pander]{pander}}.}
}
\value{
A frequency table of class \code{matrix} with added attributes used
  by \link{print} method.
}
\description{
Produces a cross-tabulation (for 2 categorical variables) with either row,
column, or total proportions as well as marginal sums.
}
\details{
The default \code{plain.ascii = TRUE} option is there to make results
  appear cleaner in the console. To avoid rmarkdown rendering problems, the
  option is automatically set to \code{FALSE} whenever
  \code{style = "rmarkdown"} (unless \code{plain.ascii = TRUE} is made
  explicit). If the intent is to produce markdown text using {style = "simple"},
  set \code{plain.ascii} to \code{FALSE}.
}
\examples{
data("tobacco")
with(tobacco, ctable(gender, smoker, prop="r"))

}
\seealso{
\code{\link[base]{table}}, \code{\link[stats]{xtabs}}
}
\author{
Dominic Comtois, \email{dominic.comtois@gmail.com}
}
\keyword{category}
\keyword{classes}
