% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descr.R
\name{descr}
\alias{descr}
\title{Univariate Statistics for Numerical Data}
\usage{
descr(x, stats = "all", na.rm = TRUE, round.digits = 2,
  transpose = FALSE, style = "simple", plain.ascii = TRUE,
  justify = "right", use.labels = FALSE, weights = NA,
  rescale.weights = FALSE, ...)
}
\arguments{
\item{x}{A numerical vector or a dataframe.}

\item{stats}{Which stats to produce. Either \dQuote{all} (default), or a
selection of : \dQuote{min}, \dQuote{median}, \dQuote{max}, \dQuote{mad},
\dQuote{iqr}, \dQuote{cv}, \dQuote{skewness}, \dQuote{se.skewness}, and
\dQuote{kurtosis}.}

\item{na.rm}{Argument to be passed to statistical functions. Defaults to
\code{TRUE}.}

\item{round.digits}{Number of significant digits to keep. Defaults to
\code{2}.}

\item{transpose}{Makes variables appears as columns, and stats as rows.
Defaults to \code{FALSE}.}

\item{style}{Style to be used by \code{\link[pander]{pander}} when rendering
output tables; One of \dQuote{simple} (default), \dQuote{grid} or
\dQuote{rmarkdown}.}

\item{plain.ascii}{Logical \code{\link[pander]{pander}} argument. When
\code{TRUE}, no markup characters will be used (useful when printing
to console). Defaults to \code{TRUE} when \code{style} is \dQuote{simple},
and \code{FALSE} otherwise.}

\item{justify}{String indicating alignment of columns; one of \dQuote{left}
(or \dQuote{l}), \dQuote{center} (or \dQuote{c}), or \dQuote{right}
(or \dQuote{r}). Defaults to \dQuote{right}.}

\item{use.labels}{Logical. Display label instead of variable name when
label exists. Available only when \code{transpose = TRUE}.}

\item{weights}{Vector of weights having same length as x. Use \code{NA}
(default) when no weights are provided.}

\item{rescale.weights}{Logical parameter. When set to \code{TRUE}, the
total count will be the same as the unweighted \code{x}. \code{FALSE} by
default.}

\item{\dots}{Additional arguments passed to \code{\link[pander]{pander}}.}
}
\value{
A list with 2 matrix elements: the stats themselves, and a summary
  of valid / missing cases. The list also has attributes used by
  \pkg{summarytool}'s print method.
}
\description{
Calculates weighted or non-weighted mean, standard deviation, min, max,
median, mad, IQR*, CV, skewness*, SE.skewness*, and kurtosis* on numerical
vectors. (Items marked with an * are only available for non-weighted
vectors / dataframes.
}
\details{
The default \code{plain.ascii = TRUE} option is there to make results
  appear cleaner in the console. To avoid rmarkdown rendering problems, the
  option is automatically set to \code{FALSE} whenever
  \code{style = "rmarkdown"} (unless \code{plain.ascii = TRUE} is made
  explicit). If the intent is to produce markdown text using {style = "simple"},
  set \code{plain.ascii} to \code{FALSE}.
}
\examples{
data(exams)
descr(exams)
descr(exams, transpose=TRUE)
descr(exams, stats = c("Mean", "Std.Dev"))
data(tobacco)
with(tobacco, by(age, smoker, descr))

}
\author{
Dominic Comtois, \email{dominic.comtois@gmail.com}
}
\keyword{univar}
