% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfSummary.R
\name{dfSummary}
\alias{dfSummary}
\title{Dataframe Summary}
\usage{
dfSummary(x, round.digits = 2, varnumbers = FALSE,
  display.labels = any(sapply(X = x, FUN = Hmisc::label) != ""),
  style = "multiline", plain.ascii = TRUE, justify = "left",
  max.distinct.values = 10, trim.strings = FALSE, max.string.width = 25,
  split.cells = 40, split.table = Inf, ...)
}
\arguments{
\item{x}{A dataframe.}

\item{round.digits}{Number of significant digits to display in numerical
summaries and in frequency proportions. Defaults to \code{2}.}

\item{varnumbers}{Should the first column contain variable number? Defaults
to \code{FALSE}.}

\item{display.labels}{If \code{TRUE}, variable labels (as defined with
\pkg{rapportools} or \pkg{Hmisc}'s \code{label} functions) will be displayed.
By default, the \emph{labels} column is shown if at least one of the columns
has a defined label.}

\item{style}{The style to be used by \code{\link[pander]{pander}} when
renderingin output table. Defaults to \dQuote{multiline}. Another option is
\dQuote{grid}. Style \dQuote{simple} is not supported for this particular
function, and \dQuote{rmarkdown} will fallback to \dQuote{multiline}.}

\item{plain.ascii}{Logical \code{\link[pander]{pander}} argument. When
\code{TRUE}, no markup characters will be used (useful when printing
to console). Defaults to \code{TRUE}.}

\item{justify}{String indicating alignment of columns; one of \dQuote{left}
(or \dQuote{l}), \dQuote{center} (or \dQuote{c}), or \dQuote{right}
(or \dQuote{r}). Defaults to \dQuote{right}.}

\item{max.distinct.values}{The maximum number of values to display frequencies
for. If variable has more distinct values than this parameter, the remaining
frequencies will be reported as a whole, along with the number of additional
distinct values.}

\item{trim.strings}{Logical; for character variables, should leading and
trailing white space be removed? When set to \code{TRUE}, this is done
\emph{before} calculating frequencies, so those will be impacted
accordingly. Interpret frequencies accordingly. Defaults to \code{FALSE}.}

\item{max.string.width}{Limits the number of characters to display in the
frequency tables. Defaults to \code{25}.}

\item{split.cells}{A numeric argument passed to \code{\link[pander]{pander}}.
It is the number of characters allowed on a line before splitting the cell.
Defaults to \code{40}.}

\item{split.table}{\pkg{pander} argument which determines the maximum
width of a table. Keeping the default value (\code{Inf}) is recommended.}

\item{\dots}{Additional arguments passed to \code{\link[pander]{pander}}.}
}
\value{
A dataframe containing as many rows as there are columns in \code{x},
  with additionnal attributes to inform \code{print} function. Columns of the
  output dataframe are:
  \describe{
    \item{Var Number}{Number indicating the order in which column appears in
      the dataframe.}
    \item{Variable}{Name of the variable.}
    \item{Label}{Label of the variable (set by using \code{Hmisc::label()} or
       \code{rapportools::label()}).}
    \item{Properties}{Type & class(es) of the variable.}
    \item{Stats / Values}{For factors, a list of their values, limited by the
      \code{max.distinct.values} parameter. For character variables, the most
      common values (in descending frequency order), also limited by
      \code{max.distinct.values}. For numerical variables, common univariate
      statistics (mean, std. deviation, min, med, max, IQR and CV).}
    \item{Freqs, \% of Valid}{For factors and character variables, the frequencies
      and proportions of the values listed in the previous column. For numerical
      vectors, number of distinct values, or frequency of distinct values if
      their number is not greater than \code{max.distinct.values}.}
    \item{N Valid}{Number and proportion of valid values for the variable.}
}
}
\description{
Summary of a dataframe consisting of: variable names, labels if any, factor
levels, frequencies and/or numerical summary statistics, and valid/missing
observations information.
}
\details{
The default \code{plain.ascii = TRUE} option is there to make results
  appear cleaner in the console. When used in a context of markdown rendering,
  set this option to \code{FALSE}.
}
\examples{
data(tobacco)
dfSummary(tobacco)
\dontrun{view(dfSummary(iris))}

}
\seealso{
\code{\link[base]{summary.data.frame}}
}
\author{
Dominic Comtois, \email{dominic.comtois@gmail.com}
}
\keyword{attribute}
\keyword{category}
\keyword{classes}
\keyword{univar}
