\name{examens}
\alias{examens}
\docType{data}
\title{
examens sample data (French)
}
\description{
This is merely a sample dataframe to allow users to experiment with summarytools. Am English  equivalent called 'exams' is also available.
}
\usage{data("examens")}
\format{
  A dataframe with 25 observations with the following 6 variables.
  \describe{
    \item{\code{francais}}{a numerical vector}
    \item{\code{maths}}{a numerical vector}
    \item{\code{geographie}}{a numerical vector}
    \item{\code{histoire}}{a numerical vector}
    \item{\code{economie}}{a numerical vector}
    \item{\code{anglais}}{a numerical vector}
  }
}
\examples{
data(examens)
descr(examens)
}
\keyword{datasets}
