% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examens.R
\docType{data}
\name{examens}
\alias{examens}
\title{Bulletin de Notes (Donne simulees)}
\format{Un data frame de 30 rangees et 8 colonnes}
\usage{
data(examens)
}
\description{
Jeu de donnees simulees contenant les notes de 30 etudiants, avec les colonnes suivantes:
\itemize{
  \item etudiant Nom de l'etudiant.
  \item sexe Variable categorielle (facteur). Deux niveaux: \dQuote{Fille}, \dQuote{Garcon}.
  \item francais Note en francais (numerique).
  \item math Note en maths (numerique).
  \item geographie Note en geographie (numerique).
  \item histoire Note en histoire (numerique).
  \item economie Note en economie (numerique).
  \item anglais Note en anglais (numerique).
}
}
\details{
Donnees simulees. Les notes de chaque etudiant sont centrees autour d'une moyenne
personnelle et ecart-type randomises.

A copy of this dataset is \strong{available in English} under the name
\dQuote{exams}.
}
\keyword{datasets}
