% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examens.R
\docType{data}
\name{examens}
\alias{examens}
\title{Bulletin de Notes (Données simulées)}
\format{Un data frame de 30 rangées et 8 colonnes}
\usage{
data(examens)
}
\description{
Jeu de données simulées contenant les notes de 30 étudiants, avec les colonnes suivantes:
\itemize{
  \item etudiant Nom de l'étudiant.
  \item sexe Variable catégorielle (facteur) à 2 niveaux: \dQuote{Fille}, \dQuote{Garçon}.
  \item francais Note en français (numérique).
  \item math Note en maths (numérique).
  \item geographie Note en géographie (numérique).
  \item histoire Note en histoire (numérique).
  \item economie Note en économie (numérique).
  \item anglais Note en anglais (numérique).
}
}
\details{
Données simulées. Les notes de chaque étudiant sont centrées autour d'une moyenne
personnelle et écart-type randomisés.

A copy of this dataset is \strong{available in English} under the name
\dQuote{exams}.
}
\keyword{datasets}
