\name{exams}
\alias{exams}
\docType{data}
\title{
exams -- Sample Data
}
\description{
A sample dataframe to allow users to experiment with summarytools. A French equivalent called 'examens' is also available.
}
\usage{data("exams")}
\format{
  A data frame with 25 observations with the following 6 variables.
  \describe{
    \item{\code{gender}}{a factor with levels "F" and "M"}
    \item{\code{english}}{a numeric vector}
    \item{\code{math}}{a numeric vector}
    \item{\code{geography}}{a numeric vector}
    \item{\code{history}}{a numeric vector}
    \item{\code{economics}}{a numeric vector}
    \item{\code{french}}{a numeric vector}
  }
}
\keyword{datasets}
