% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exams.R
\docType{data}
\name{exams}
\alias{exams}
\title{Report Cards - Simulated Data}
\format{A data frame with 30 rows and 8 variables}
\usage{
data(exams)
}
\description{
A simulated dataset with grades for hypothetical 30 students, with
the following variables:
\itemize{
  \item student Student's name.
  \item gender Factor with 2 levels: \dQuote{Girl}, \dQuote{Boy}.
  \item french French Grade (numerical).
  \item math Math Grade (numerical).
  \item geography Geography Grade (numerical).
  \item history History Grade (numerical).
  \item economics Economics Grade (numerical).
  \item english English Grade (numerical).
}
}
\details{
All names and grades are simulated. Grades for each student are centered
around a personal randomized average and standard deviation.

Une version française de ce \emph{data frame} est disponible sous le nom 
d'\dQuote{examens}.
}
\keyword{datasets}
