% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq.R
\name{freq}
\alias{freq}
\title{Frequency Tables for Factors and Other Discrete Data}
\usage{
freq(
  x,
  var = NULL,
  round.digits = st_options("round.digits"),
  order = "default",
  style = st_options("style"),
  plain.ascii = st_options("plain.ascii"),
  justify = "default",
  cumul = st_options("freq.cumul"),
  totals = st_options("freq.totals"),
  report.nas = st_options("freq.report.nas"),
  rows = numeric(),
  missing = "",
  na.val = st_options("na.val"),
  display.type = TRUE,
  display.labels = st_options("display.labels"),
  headings = st_options("headings"),
  weights = NA,
  rescale.weights = FALSE,
  ...
)
}
\arguments{
\item{x}{Factor, vector, or data frame.}

\item{var}{Optional unquoted variable name. Provides support for piped
function calls (e.g. \code{my_df \%>\% freq(my_var)}).}

\item{round.digits}{Numeric. Number of significant digits to display. 
Defaults to \code{2}. Can be set globally with \code{\link{st_options}}.}

\item{order}{Character. Ordering of rows in frequency table; \dQuote{name}
(default for non-factors), \dQuote{level} (default for factors), or
\dQuote{freq} (from most frequent to less frequent). To invert the order,
place a minus sign before or after the word. \dQuote{-freq} will thus
display the items starting from the lowest in frequency to the highest,
and so forth.}

\item{style}{Character. Style to be used by \code{\link[pander]{pander}}. One
of \dQuote{simple} (default), \dQuote{grid}, \dQuote{rmarkdown}, or
\dQuote{jira}. Can be set globally with \code{\link{st_options}}.}

\item{plain.ascii}{Logical. \code{\link[pander]{pander}} argument; when
\code{TRUE}, no markup characters will be used (useful when printing to
console). Defaults to \code{TRUE} unless \code{style = 'rmarkdown'}, in
which case it will be set to \code{FALSE} automatically. Can be set
globally with \code{\link{st_options}}.}

\item{justify}{String indicating alignment of columns. By default
(\dQuote{default}), \dQuote{right} is used for text tables and
\dQuote{center} is used for \emph{html} tables. You can force it to one of
\dQuote{left}, \dQuote{center}, or \dQuote{right}.}

\item{cumul}{Logical. Set to \code{FALSE} to hide cumulative proportions
from results. \code{TRUE} by default. To change this value globally, see 
 \code{\link{st_options}}.}

\item{totals}{Logical. Set to \code{FALSE} to hide totals from results. 
\code{TRUE} by default. To change this value globally, see 
\code{\link{st_options}}.}

\item{report.nas}{Logical. Set to \code{FALSE} to turn off reporting of
missing values. To change this default value globally, see
\code{\link{st_options}}.}

\item{rows}{Character or numeric vector allowing subsetting of the results.
The order given here will be reflected in the resulting table. If a single
string is used, it will be used as a regular expression to filter row 
names.}

\item{missing}{Text to display in NA cells. Defaults to \dQuote{}.}

\item{na.val}{Character. For factors and character vectors, consider this
value as \code{NA}. Ignored if there are actual NA values or if it matches
no value / factor level in the data. \code{NULL} by default.}

\item{display.type}{Logical. Should variable type be displayed? Default is
\code{TRUE}.}

\item{display.labels}{Logical. Should variable / data frame labels be
displayed? Default is \code{TRUE}. To change this default value globally,
see \code{\link{st_options}}.}

\item{headings}{Logical. Set to \code{FALSE} to omit heading section. Can be
set globally via \code{\link{st_options}}.}

\item{weights}{Vector of weights; must be of the same length as \code{x}.}

\item{rescale.weights}{Logical parameter. When set to \code{TRUE}, the total
count will be the same as the unweighted \code{x}. \code{FALSE} by default.}

\item{\dots}{Additional arguments passed to \code{\link[pander]{pander}}.}
}
\value{
A frequency table of class \code{matrix} and \code{summarytools} with
  added attributes used by \emph{print} method.
}
\description{
Displays weighted or unweighted frequencies, including <NA> counts and
proportions.
}
\details{
The default \code{plain.ascii = TRUE} option is there to make
  results appear cleaner in the console. To avoid rmarkdown rendering
  problems, this option is automatically set to \code{FALSE} whenever
  \code{style = "rmarkdown"} (unless \code{plain.ascii = TRUE} is made
  explicit in the function call).
}
\note{
The data type represents the \code{\link[base]{class}} in most cases.
}
\examples{
data(tobacco)
freq(tobacco$gender)
freq(tobacco$gender, totals = FALSE)

# Ignore NA's, don't show totals, omit headings
freq(tobacco$gender, report.nas = FALSE, totals = FALSE, headings = FALSE)

# In .Rmd documents, use the two following arguments, minimally
freq(tobacco$gender, style="rmarkdown", plain.ascii = FALSE)

# Grouped Frequencies
with(tobacco, stby(diseased, smoker, freq))
(fr_smoker_by_gender <- with(tobacco, stby(smoker, gender, freq)))

# Print html Source
print(fr_smoker_by_gender, method = "render", footnote = NA)

# Order by frequency (+ to -)
freq(tobacco$age.gr, order = "freq")

# Order by frequency (- to +)
freq(tobacco$age.gr, order = "-freq")

# Use the 'rows' argument to display only the 10 most common items
freq(tobacco$age.gr, order = "freq", rows = 1:10)

\dontrun{
# Display rendered html results in RStudio's Viewer
# notice 'view()' is NOT written with capital V
# If working outside RStudio, Web browser is used instead
# A temporary file is stored in temp dir
view(fr_smoker_by_gender)

# Display rendered html results in default Web browser
# A temporary file is stored in temp dir here too
print(fr_smoker_by_gender, method = "browser")

# Write results to text file (.txt, .md, .Rmd) or html file (.html)
print(fr_smoker_by_gender, method = "render", file = "fr_smoker_by_gender.md)
print(fr_smoker_by_gender, method = "render", file = "fr_smoker_by_gender.html)
}

}
\seealso{
\code{\link[base]{table}}
}
\author{
Dominic Comtois, \email{dominic.comtois@gmail.com}
}
\keyword{category}
\keyword{classes}
\keyword{univar}
