% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq.R
\name{freq}
\alias{freq}
\title{Frequency Tables for Factors and Other Discrete Data}
\usage{
freq(x, round.digits = st_options("round.digits"), order = "default",
  style = st_options("style"), plain.ascii = st_options("plain.ascii"),
  justify = "default", cumul = st_options("freq.cumul"),
  totals = st_options("freq.totals"),
  report.nas = st_options("freq.report.nas"), missing = "",
  display.type = TRUE, display.labels = st_options("display.labels"),
  headings = st_options("headings"), weights = NA,
  rescale.weights = FALSE, ...)
}
\arguments{
\item{x}{Factor or vector}

\item{round.digits}{Number of significant digits to display. Defaults to
\code{2} and can be set globally; see \code{\link{st_options}}.}

\item{order}{Ordering of rows in frequency table; \dQuote{names} (default for
non-factors), \dQuote{levels} (default for factors), or \dQuote{freq} (from
most frequent to less frequent). Can also be a numeric or character vector
specifying the desired order for several or all frequency table rows.}

\item{style}{Style to be used by \code{\link[pander]{pander}} when rendering
output table; One of \dQuote{simple} (default), \dQuote{grid}, or
\dQuote{rmarkdown} This option can be set globally; see
\code{\link{st_options}}.}

\item{plain.ascii}{Logical. \code{\link[pander]{pander}} argument; when
\code{TRUE}, no markup characters will be used (useful when printing to
console). Defaults to \code{TRUE} unless \code{style = 'rmarkdown'}, in
which case it will be set to \code{FALSE} automatically. To change the
default value globally, see \code{\link{st_options}}.}

\item{justify}{String indicating alignment of columns. By default
(\dQuote{default}), \dQuote{right} is used for text tables and
\dQuote{center} is used for \emph{html} tables. You can force it to one of
\dQuote{left}, \dQuote{center}, or \dQuote{right}.}

\item{cumul}{Logical. Set to \code{FALSE} to hide cumulative proportions
from results. \code{TRUE} by default. To change this value globally, see 
 \code{\link{st_options}}.}

\item{totals}{Logical. Set to \code{FALSE} to hide totals from results. 
\code{TRUE} by default. To change this value globally, see 
\code{\link{st_options}}.}

\item{report.nas}{Logical. Set to \code{FALSE} to turn off reporting of
missing values. To change this default value globally, see
\code{\link{st_options}}.}

\item{missing}{Characters to display in NA cells. Defaults to \dQuote{}.}

\item{display.type}{Logical. Should variable type be displayed? Default is
\code{TRUE}.}

\item{display.labels}{Logical. Should variable / data frame labels be
displayed? Default is \code{TRUE}. To change this default value globally,
see \code{\link{st_options}}.}

\item{headings}{Logical. Set to \code{FALSE} to omit heading section. Can be
set globally via \code{\link{st_options}}.}

\item{weights}{Vector of weights; must be of the same length as \code{x}.}

\item{rescale.weights}{Logical parameter. When set to \code{TRUE}, the total
count will be the same as the unweighted \code{x}. \code{FALSE} by default.}

\item{\dots}{Additional arguments passed to \code{\link[pander]{pander}}.}
}
\value{
A frequency table of class \code{matrix} and \code{summarytools} with
  added attributes used by \emph{print} method.
}
\description{
Displays weighted or unweighted frequencies, including <NA> counts and
proportions.
}
\details{
The default \code{plain.ascii = TRUE} option is there to make
  results appear cleaner in the console. To avoid rmarkdown rendering
  problems, this option is automatically set to \code{FALSE} whenever
  \code{style = "rmarkdown"} (unless \code{plain.ascii = TRUE} is made
  explicit in the function call).
}
\examples{
data(tobacco)
freq(tobacco$gender)
freq(tobacco$gender, totals = FALSE)

# Ignore NA's, don't show totals, omit headings
freq(tobacco$gender, report.nas = FALSE, totals = FALSE, headings = FALSE)

# In .Rmd documents, use the two following arguments, minimally
freq(tobacco$gender, style="rmarkdown", plain.ascii = FALSE)

# Grouped Frequencies
with(tobacco, stby(diseased, smoker, freq))
(fr_smoker_by_gender <- with(tobacco, stby(smoker, gender, freq)))

# Print html Source
print(fr_smoker_by_gender, method = "render", footnote = NA)

\dontrun{
# Display rendered html results in RStudio's Viewer
# notice 'view()' is NOT written with capital V
# If working outside RStudio, Web browser is used instead
# A temporary file is stored in temp dir
view(fr_smoker_by_gender)

# Display rendered html results in default Web browser
# A temporary file is stored in temp dir here too
print(fr_smoker_by_gender, method = "browser")

# Write results to text file (.txt, .md, .Rmd) or html file (.html)
print(fr_smoker_by_gender, method = "render", file = "fr_smoker_by_gender.md)
print(fr_smoker_by_gender, method = "render", file = "fr_smoker_by_gender.html)
}

}
\seealso{
\code{\link[base]{table}}
}
\author{
Dominic Comtois, \email{dominic.comtois@gmail.com}
}
\keyword{category}
\keyword{classes}
\keyword{univar}
