% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numericSummary.R
\name{numericSummary}
\alias{numericSummary}
\title{Numeric Summary}
\usage{
numericSummary(vec, na.rm = TRUE)
}
\arguments{
\item{vec}{A numeric vector of length > 1}

\item{na.rm}{Logical. Remove \code{NA} before calculating summaries}
}
\value{
Named vector of summary statistics
}
\description{
Numeric Summary
}
\examples{
numericSummary(airquality$Wind)
}

