% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oneYearOneLocSummary.R
\name{oneYearOneLocSummary}
\alias{oneYearOneLocSummary}
\title{oneYearOneLocSummary function}
\usage{
oneYearOneLocSummary(dF, traits, sortby = NULL, allowDupEnt = TRUE,
  unitSep = "|", ...)
}
\arguments{
\item{dF}{[value]}

\item{traits}{[value]}

\item{sortby}{[value]. Default is NULL}

\item{allowDupEnt}{[value]. Default is TRUE}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
none
}
