% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oneYearOneLocSummary.R
\name{oneYearOneLocSummary}
\alias{oneYearOneLocSummary}
\title{oneYearOneLocSummary function}
\usage{
oneYearOneLocSummary(
  dF,
  traits,
  sortHiLo = NULL,
  sortLoHi = NULL,
  allowDupEnt = TRUE,
  unitSep = "|",
  verbose = FALSE,
  fixed = NULL,
  random = NULL,
  printFit = FALSE,
  ...
)
}
\arguments{
\item{dF}{[value]}

\item{traits}{[value]}

\item{allowDupEnt}{[value]. Default is TRUE}

\item{sortby}{[value]. Default is NULL}
}
\value{
[value]
}
\description{
function to (do something)
}
\details{
[fill in details here]
}
\examples{
# none
}
