\name{parse_arg}
\alias{parse_arg}
\title{
Extract Information About Data From x Argument
}
\description{
Used internally to extract information on the data being described. 
}
\usage{
.parse_arg(scalls, sframes, mcall)
}
\arguments{
  \item{scalls}{Object returned by \code{sys.calls()}}
  \item{sframes}{Object returned by \code{sys.frames()}}
  \item{mcall}{Object returned by \code{match.call()}}
}
\details{
The extracted attributes are attached to the objects generated by \code{freq()}, \code{descr()}
or \code{dfSummary()}.
}
\value{
A list comprised of:
\item{df.name}{The dataframe name when applicable}
\item{var.names}{The variable names when applicable}
\item{rows.subset}{The subsetting condition when applicable}
\item{by.group}{The group, when functions are called through \code{by()}}
}
\references{\url{https://github.com/dcomtois/summarytools}}
\author{Dominic Comtois <dominic.comtois@gmail.com>}
\keyword{ internal }
\keyword{ misc }
