\name{parse_arg}
\alias{parse_arg}
\title{
Extract Information About Data From x Argument
}
\description{
Used internally to extract information on the data being described.
}
\usage{
.parse_arg(sys_calls, sys_frames, match_call)
}
\arguments{
  \item{sys_calls}{Object returned by \code{sys.calls()}}
  \item{sys_frames}{Object returned by \code{sys.frames()}}
  \item{match_call}{Object returned by \code{match.call()}}
}
\details{
The extracted attributes are attached to the objects generated by \code{freq()}, \code{descr()}
or \code{dfSummary()}.
}
\value{
A list comprised of:
\item{df_name}{The dataframe name when applicable}
\item{var_names}{The variable names when applicable}
\item{rows_subset}{The subsetting condition when applicable}
\item{by_group}{The group, when functions are called through \code{by()}}
}
\references{\url{https://github.com/dcomtois/summarytools}}
\author{Dominic Comtois <dominic.comtois@gmail.com>}
\keyword{ internal }
\keyword{ misc }
