% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{st_options}
\alias{st_options}
\title{Displaying and setting summarytools global options}
\usage{
st_options(option, value)
}
\arguments{
\item{option}{option name (string).}

\item{value}{value to assign (optional)}
}
\description{
To list all \code{summarytools} global options, run this function without any parameters. 
To display the value of an option, use the first parameter only. To modify it, add the new
value as a second parameter.
}
\details{
The following options are available:

\itemize{
  \item \code{round.digits}  Numeric. Default: \code{2}.
  \item \code{plain.ascii}  Logical. \code{TRUE} by default. Set to \code{FALSE} when using
    summarytools with a rendering tool such as \code{knitr} or when creating output files
    to be converted with Pandoc.
  \item \code{omit.headings}  Logical. \code{FALSE} by default. Set to \code{TRUE} to remove all
    headings from the outputs (only the tables will be printed out).
  \item \code{footnote}  Character. When the default value 'default' is used, the package name,
    version, and R + version are displayed below html outputs. Set no \code{NA} to omit
    the footnote, or provide a string to personalize it.
  \item \code{display.labels}  Logical. \code{TRUE} by default. Set to \code{FALSE} to omit data 
    frame and variable labels in the headings section.
  \item \code{freq.totals}  Logical. Corresponds to the \code{totals} parameter of \code{\link{freq}}.
    \code{TRUE} by default.
  \item \code{freq.display.nas}  Logical. Corresponds to the \code{display.nas} parameter of 
    \code{freq()}. \code{TRUE} by default.
  \item \code{ctable.totals}  Logical. Corresponds to the \code{totals} parameter of 
    \code{\link{ctable}}. \code{TRUE} by default.
  \item \code{ctable.prop}  Logical. Corresponds to the \code{totals} parameter of 
    \code{\link{ctable}}.
  \item \code{descr.transpose}  Logical. Corresponds to the \code{transpose} parameter of 
    \code{\link{descr}}. \code{FALSE} by default.
  \item \code{bootstrap.css}  Logical.  Include Bootstrap CSS in html outputs. Defaults to 
    \code{TRUE}. When using the \dQuote{render} method, it may be a good idea is to set this
    to \code{FALSE}.
  \item \code{custom.css}  Character. Path to an additional, user-provided CSS file. \code{NA} 
    by default.
  \item \code{escape.pipe}  Logical. \code{FALSE} by default. Set to \code{TRUE} if Pandoc conversion
    is your goal and you have unsatisfying results with grid or multiline tables.
    
}
}
\examples{
\dontrun{
st_options()
st_options('round.digits')
st_options('round.digits', 1)
}
}
