% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryTools-package.R
\docType{package}
\name{summaryTools}
\alias{summaryTools}
\alias{summaryTools-package}
\title{Summary Tools}
\description{
This package has functions for summarising common R data types
}
\author{
Doug Ashton
}

