\name{tabagisme}
\alias{tabagisme}
\docType{data}
\title{
  tabagisme -- Sample Data (French)
}
\description{
This is merely a sample dataframe to allow users to experiment with summarytools. An english equivalent called 'tobacco' is also available.
}
\usage{data("tabagisme")}
\format{
  A data frame with 300 observations on the following 5 variables.
  \describe{
    \item{\code{sexe}}{Factor with 2 levels, 'F' and 'M'}
    \item{\code{age}}{Numerical vector}
    \item{\code{age.gr}}{Factor with 5 levels (age categories)}
    \item{\code{IMC}}{Numerical vector (Indice de masse corporelle / Body Mass Index)}
    \item{\code{fumeur}}{Factor with 2 levels (smoking status)}
    \item{\code{malade}}{Factor with 2 levels (diseased)}

  }
}
\keyword{datasets}
