% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label.R
\name{unlabel}
\alias{unlabel}
\title{Clear Variable and Data Frame Label(s)}
\usage{
unlabel(x)
}
\arguments{
\item{x}{An R object to remove labels from.}
}
\description{
Returns the object with all labels removed. The \dQuote{label} attribute
as well as the \dQuote{labelled} class (used by Hmisc and labelled) are
cleared.
}
\seealso{
\code{\link{label}}
}
\author{
Dominic Comtois, \email{dominic.comtois@gmail.com},
}
