% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label.R
\name{unlabel}
\alias{unlabel}
\title{Clear Variable or Data Frame Label(s)}
\usage{
unlabel(x)
}
\arguments{
\item{x}{An R object to remove labels from.}
}
\description{
Returns the object with all labels removed. Both the \dQuote{label} attribute
and \pkg{Hmisc}'s \dQuote{labelled} class are removed.
}
\author{
Dominic Comtois, \email{dominic.comtois@gmail.com},
}
