% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useTranslations.R
\name{useTranslations}
\alias{useTranslations}
\title{Import and use a custom language}
\usage{
useTranslations(file)
}
\arguments{
\item{file}{Character. The path to the translations file.}
}
\description{
If your language is not available or if you wish to customize the outputs'
language to suit your preference, you can set up a translations file (see
details) and import it with this function.
}
\details{
To build the translations file, you can get
\href{https://raw.githubusercontent.com/dcomtois/summarytools/dev-current/translations/custom_lang_template.csv}{this template},
or use the \href{https://raw.githubusercontent.com/dcomtois/summarytools/dev-current/translations/translations.csv}{current translations definition file}
and fill out a new column \strong{having \sQuote{custom} on the first line}.
The file must have \emph{UTF-8} encoding.
}
\keyword{utilities}
